+++ CHANGELOG.md
@@ -10,7 +10,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ### Added
 
-- Add `default` keyword arg to `env()`
+- Add `default` keyword arg to `env()` [#652](https://github.com/LucasPickering/slumber/issues/652)
+- Add string maniuplation functions [#655](https://github.com/LucasPickering/slumber/issues/655)
+  - `split()` splits a string on a delimiter
 
 ## [4.2.1] - 2025-11-26